feat: move SQL template functions into new migration box
