--- pkgerx/testdata/0_sql_create_tablename_template.expected.sql
@@ -1,1 +1,0 @@
-CREATE TABLE test_table_name ( "id" UUID NOT NULL, PRIMARY KEY ("id"));
