--- pkgerx/testdata/0_sql_create_tablename_template.up.sql
@@ -1,1 +1,0 @@
-CREATE TABLE {{ identifier .Parameters.tableName }} ( "id" UUID NOT NULL, PRIMARY KEY ("id"));
