+++ popx/sql_template_funcs.go
@@ -1,4 +1,4 @@
-package pkgerx
+package popx
 
 import (
 	"fmt"
